Begin implementing resources

+++ src/Folklore/Panneau/Http/Controllers/ResourceController.php
@@ -5,25 +5,67 @@ namespace Folklore\Panneau\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
+use Folklore\Panneau\Support\Resource;
 
-abstract class ResourceController extends Controller
+class ResourceController extends Controller
 {
     protected $indexPaginated = false;
     protected $resultsPerPage = 25;
     protected $pageInputName = 'page';
     protected $returnPagination = true;
+    protected $resourceParamName = 'resource';
+    protected $idParamName = 'id';
 
-    abstract protected function getResourceClass();
+    public function __construct()
+    {
+        $this->resourceParamName = config('panneau.route_resource_param', 'resource');
+        $this->idParamName = config('panneau.route_id_param', 'id');
+    }
+
+    protected function getResourceClass($resourceName)
+    {
+        $resource = app('panneau')->resource($resourceName);
+        return $resource;
+    }
 
     /**
      * Get the resource model
      *
      * @return \Illuminate\Database\Eloquent\Model
      */
-    protected function getResourceModel()
+    protected function getResourceModel($resourceName)
     {
-        $class = $this->getResourceClass();
-        return app($class);
+        $class = $this->getResourceClass($resourceName);
+        if (!is_null($class)) {
+            $model = $class->getModel();
+            if (!is_null($model)) {
+                return app($model);
+            }
+        }
+        return null;
+    }
+
+    protected function getResourceController($resourceName)
+    {
+        $class = $this->getResourceClass($resourceName);
+        if (!is_null($class)) {
+            return $class->getController();
+        }
+        return null;
+    }
+
+    /**
+     * Get the resource definition
+     *
+     * @return array
+     */
+    protected function getResourceDefinition($resourceName)
+    {
+        $class = $this->getResourceClass($resourceName);
+        if (!is_null($class)) {
+            return $class->toArray();
+        }
+        return null;
     }
 
     /**
@@ -31,9 +73,10 @@ abstract class ResourceController extends Controller
      *
      * @return \Illuminate\Database\Eloquent\Builder
      */
-    protected function getResourceQueryBuilder()
+    protected function getResourceQueryBuilder(Request $request)
     {
-        $model = $this->getResourceModel();
+        $resource = $request->resource;
+        $model = $this->getResourceModel($resource);
         return $model->newQuery();
     }
 
@@ -74,7 +117,7 @@ abstract class ResourceController extends Controller
 
     protected function getItems(Request $request)
     {
-        $query = $this->getResourceQueryBuilder();
+        $query = $this->getResourceQueryBuilder($request);
 
         if (method_exists($this, 'buildQueryFromRequest')) {
             $this->buildQueryFromRequest($query, $request);
@@ -94,8 +137,27 @@ abstract class ResourceController extends Controller
         if ($id instanceof Model) {
             $id = $id->getKey();
         }
-        $query = $this->getResourceQueryBuilder();
+        $query = $this->getResourceQueryBuilder($request);
         return $query->where('id', $id)->first();
+    }
+
+    /**
+     * Return the definition of a resource
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function definition(Request $request, $resource)
+    {
+        $customController = $this->getResourceController($resource);
+        if (!is_null($customController)) {
+            return redirect()->action($customController.'@definition', [
+                'resource' => $resource,
+            ]);
+        }
+
+        $definition = $this->getResourceDefinition($resource);
+        return $definition;
     }
 
     /**
@@ -104,8 +166,15 @@ abstract class ResourceController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function index(Request $request)
+    public function index(Request $request, $resource)
     {
+        $customController = $this->getResourceController($resource);
+        if (!is_null($customController)) {
+            return redirect()->action($customController.'@index', [
+                $this->resourceParamName => $resource,
+            ]);
+        }
+
         $items = $this->getItems($request);
 
         if ($request->wantsJson()) {
@@ -125,8 +194,15 @@ abstract class ResourceController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function create(Request $request)
+    public function create(Request $request, $resource)
     {
+        $customController = $this->getResourceController($resource);
+        if (!is_null($customController)) {
+            return redirect()->action($customController.'@create', [
+                $this->resourceParamName => $resource,
+            ]);
+        }
+
         return $this->getResourceView('create');
     }
 
@@ -136,11 +212,18 @@ abstract class ResourceController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(Request $request, $resource)
     {
+        $customController = $this->getResourceController($resource);
+        if (!is_null($customController)) {
+            return redirect()->action($customController.'@store', [
+                $this->resourceParamName => $resource,
+            ]);
+        }
+
         $data = $this->getStoreDataFromRequest($request);
 
-        $model = $this->getResourceModel();
+        $model = $this->getResourceModel($request);
         $model->fill($data);
         $model->save();
 
@@ -158,8 +241,16 @@ abstract class ResourceController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show(Request $request, $id)
+    public function show(Request $request, $resource, $id)
     {
+        $customController = $this->getResourceController($resource);
+        if (!is_null($customController)) {
+            return redirect()->action($customController.'@show', [
+                $this->resourceParamName => $resource,
+                $this->idParamName => $id,
+            ]);
+        }
+
         $item = $this->getItem($id, $request);
 
         if ($request->wantsJson()) {
@@ -178,8 +269,16 @@ abstract class ResourceController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit(Request $request, $id)
+    public function edit(Request $request, $resource, $id)
     {
+        $customController = $this->getResourceController($resource);
+        if (!is_null($customController)) {
+            return redirect()->action($customController.'@edit', [
+                $this->resourceParamName=> $resource,
+                $this->idParamName => $id,
+            ]);
+        }
+
         $item = $this->getItem($id, $request);
 
         if ($request->wantsJson()) {
@@ -198,8 +297,16 @@ abstract class ResourceController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, $resource, $id)
     {
+        $customController = $this->getResourceController($resource);
+        if (!is_null($customController)) {
+            return redirect()->action($customController.'@update', [
+                $this->resourceParamName => $resource,
+                $this->idParamName => $id,
+            ]);
+        }
+
         $data = $this->getUpdateDataFromRequest($request);
 
         $model = $this->getItem($id, $request);
@@ -220,8 +327,16 @@ abstract class ResourceController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Request $request, $id)
+    public function destroy(Request $request, $resource, $id)
     {
+        $customController = $this->getResourceController($resource);
+        if (!is_null($customController)) {
+            return redirect()->action($customController.'@destroy', [
+                $this->resourceParamName => $resource,
+                $this->idParamName => $id,
+            ]);
+        }
+
         $model = $this->getItem($id, $request);
         $data = $model->toArray();
         $model->delete();
